add connectionLogic and start parsing message

+++ javaLab3/src/main/java/org/example/connection/MessageIdE.java
@@ -1,23 +1,23 @@
 package org.example.connection;
 
 public enum MessageIdE {
-    CHOKE((byte) 0),
-    UNCHOKE((byte) 1),
-    INTERESTED((byte) 2),
-    NOT_INTERESTED((byte) 3),
-    HAVE((byte) 4),
-    BITFIELD((byte) 5),
-    REQUEST((byte) 6),
-    PIECE((byte) 7),
-    CANCEL((byte) 8);
+    CHOKE(0),
+    UNCHOKE(1),
+    INTERESTED(2),
+    NOT_INTERESTED( 3),
+    HAVE(4),
+    BITFIELD(5),
+    REQUEST(6),
+    PIECE(7),
+    CANCEL(8);
 
-    private final byte value;
+    private final int value;
 
-    MessageIdE(byte value) {
+    MessageIdE(int value) {
         this.value = value;
     }
 
-    public byte getValue() {
+    public int getValue() {
         return value;
     }
 }
